+++ crates/application/src/ports/organization_repository.rs
@@ -30,8 +30,8 @@ pub trait OrganizationRepository: Send + Sync {
     async fn find_paginated<'a, E>(
         &self,
         executor: E,
-        page: u64,
-        page_size: u64,
+        page: u32,
+        page_size: u32,
     ) -> Result<(Vec<Organization>, PaginationMeta), AppError>
     where
         E: sqlx::Acquire<'a, Database = sqlx::Postgres> + Send;
